% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhyloprofile.R
\name{runPhyloProfileRibi}
\alias{runPhyloProfileRibi}
\title{Run PhyloProfile app}
\usage{
runPhyloProfileRibi()
}
\value{
A shiny application - GUI version of PhyloProfile
}
\description{
Run PhyloProfile app
}
